+++ core/notary/snapshot.go
@@ -127,6 +127,10 @@ func (n *SnapshotNotary) OfferSignatures(
 		if k.kind != kind {
 			continue
 		}
+		if v.signature != nil {
+			continue
+		}
+
 		if signature := f(k.id); signature != nil {
 			v.signature = signature
 		}
@@ -188,6 +192,7 @@ func (n *SnapshotNotary) restoreNotary(notary *types.Notary, p *types.Payload) e
 		selfSigned  = map[idKind]bool{}
 		self        = n.top.SelfVegaPubKey()
 	)
+
 	for _, s := range notary.Sigs {
 		idK := idKind{id: s.ID, kind: v1.NodeSignatureKind(s.Kind)}