+++ core/validators/signatures.go
@@ -16,6 +16,7 @@ import (
 	"context"
 	"encoding/hex"
 	"errors"
+	"fmt"
 	"sort"
 	"time"
 
@@ -46,6 +47,7 @@ type Signatures interface {
 	ClearStaleSignatures()
 	SerialisePendingSignatures() *snapshot.ToplogySignatures
 	RestorePendingSignatures(*snapshot.ToplogySignatures)
+	OfferSignatures()
 }
 
 type signatureData struct {
@@ -54,6 +56,16 @@ type signatureData struct {
 	Nonce      *num.Uint
 	EpochSeq   uint64
 	Added      bool
+}
+
+type issuedSignature struct {
+	EthAddress       string
+	SubmitterAddress string
+}
+
+type signatureWithSubmitter struct {
+	signatureData
+	SubmitterAddress string
 }
 
 type ERC20Signatures struct {
@@ -67,7 +79,7 @@ type ERC20Signatures struct {
 
 	// stored nonce's etc. to be able to generate signatures to remove/add an ethereum address from the multisig bundle
 	pendingSignatures map[string]*signatureData
-	issuedSignatures  map[string]struct{}
+	issuedSignatures  map[string]issuedSignature
 }
 
 func NewSignatures(
@@ -86,7 +98,7 @@ func NewSignatures(
 		broker:            broker,
 		isValidatorSetup:  isValidatorSetup,
 		pendingSignatures: map[string]*signatureData{},
-		issuedSignatures:  map[string]struct{}{},
+		issuedSignatures:  map[string]issuedSignature{},
 	}
 	if isValidatorSetup {
 		s.multisig = bridges.NewERC20MultiSigControl(nw.GetEthereum())
@@ -104,6 +116,74 @@ type NodeIDAddress struct {
 	NodeID           string
 	EthAddress       string
 	SubmitterAddress string
+}
+
+func (s *ERC20Signatures) OfferSignatures() {
+	s.notary.OfferSignatures(types.NodeSignatureKindERC20MultiSigSignerAdded, s.offerValidatorAddedSignatures)
+	s.notary.OfferSignatures(types.NodeSignatureKindERC20MultiSigSignerRemoved, s.offerValidatorRemovedSignatures)
+}
+
+func (s *ERC20Signatures) getSignatureWithSubmitterByResID(resID string) (*signatureWithSubmitter, error) {
+	is, ok := s.issuedSignatures[resID]
+	if !ok {
+		return nil, fmt.Errorf("unable to find issued signature with resource id %q", resID)
+	}
+
+	sd, ok := s.pendingSignatures[is.EthAddress]
+	if !ok {
+		return nil, fmt.Errorf("unable to find pending signature by ethereum address %q", is.EthAddress)
+	}
+
+	return &signatureWithSubmitter{
+		signatureData:    *sd,
+		SubmitterAddress: is.SubmitterAddress,
+	}, nil
+}
+
+func (s *ERC20Signatures) offerValidatorAddedSignatures(resID string) []byte {
+	if !s.isValidatorSetup {
+		return nil
+	}
+
+	sig, err := s.getSignatureWithSubmitterByResID(resID)
+	if err != nil {
+		s.log.Panic("unable to find signature", logging.Error(err))
+	}
+
+	if !sig.Added {
+		s.log.Panic("expected added signature but got removed signature instead", logging.String("ethereumAddress", sig.EthAddress))
+	}
+
+	signature, err := s.multisig.AddSigner(sig.EthAddress, sig.SubmitterAddress, sig.Nonce.Clone())
+	if err != nil {
+		s.log.Panic("could not sign remove signer event, wallet not configured properly",
+			logging.Error(err))
+	}
+
+	return signature.Signature.Bytes()
+}
+
+func (s *ERC20Signatures) offerValidatorRemovedSignatures(resID string) []byte {
+	if !s.isValidatorSetup {
+		return nil
+	}
+
+	sig, err := s.getSignatureWithSubmitterByResID(resID)
+	if err != nil {
+		s.log.Panic("unable to find signature", logging.Error(err))
+	}
+
+	if sig.Added {
+		s.log.Panic("expected removed signature but got added signature instead", logging.String("ethereumAddress", sig.EthAddress))
+	}
+
+	signature, err := s.multisig.RemoveSigner(sig.EthAddress, sig.SubmitterAddress, sig.Nonce.Clone())
+	if err != nil {
+		s.log.Panic("could not sign remove signer event, wallet not configured properly",
+			logging.Error(err))
+	}
+
+	return signature.Signature.Bytes()
 }
 
 func (s *ERC20Signatures) getSignatureData(nodeID string, added bool) []*signatureData {
@@ -254,7 +334,10 @@ func (s *ERC20Signatures) EmitValidatorAddedSignatures(ctx context.Context, subm
 		))
 
 		// store that we issued it for this submitter
-		s.issuedSignatures[resid] = struct{}{}
+		s.issuedSignatures[resid] = issuedSignature{
+			EthAddress:       pending.EthAddress,
+			SubmitterAddress: submitter,
+		}
 	}
 	s.broker.SendBatch(evts)
 	return nil
@@ -312,7 +395,10 @@ func (s *ERC20Signatures) EmitValidatorRemovedSignatures(ctx context.Context, su
 		))
 
 		// store that we issued it for this submitter
-		s.issuedSignatures[resid] = struct{}{}
+		s.issuedSignatures[resid] = issuedSignature{
+			EthAddress:       pending.EthAddress,
+			SubmitterAddress: submitter,
+		}
 	}
 	s.broker.SendBatch(evts)
 	return nil
@@ -374,5 +460,9 @@ func (n *noopSignatures) SerialisePendingSignatures() *snapshot.ToplogySignature
 }
 
 func (n *noopSignatures) RestorePendingSignatures(*snapshot.ToplogySignatures) {
+	n.log.Error("noopSignatures implementation in use in production")
+}
+
+func (n *noopSignatures) OfferSignatures() {
 	n.log.Error("noopSignatures implementation in use in production")
 }